fix(transaction/core): u128 wealth casts so cluster_fee tests compile under --all-features
The cluster_fee module is gated behind the non-default `cluster-tax`
feature, so default CI never compiles it. The fee methods in
cluster-tax/src/fee_curve.rs take `cluster_wealth: u128`, but the
cluster-wealth helpers in cluster_fee.rs still return u64, so four
test-side call sites failed with E0308 under
`cargo check -p bth-transaction-core --all-targets --all-features`.

Apply surgical `as u128` casts at the four compute_fee/
compute_fee_with_outputs/compute_fee_with_dynamic_base_and_outputs call
sites (lines 691, 781, 931, 1090). Deliberately do NOT widen the shared
locals or helper return types: at :781 `effective_wealth` also flows into
the u64 `InsufficientClusterFee.cluster_wealth` error field, so widening
would just push the mismatch onto that field.

Closes #1057

+++ transaction/core/src/validation/cluster_fee.rs
@@ -688,7 +688,8 @@ mod tests {
         // Calculate minimum fee
         let effective_wealth =
             compute_effective_cluster_wealth_from_tags(&input_tags, &input_values, &wealth_map);
-        let min_fee = fee_config.compute_fee(TransactionType::Hidden, 4000, effective_wealth, 0);
+        let min_fee =
+            fee_config.compute_fee(TransactionType::Hidden, 4000, effective_wealth as u128, 0);
 
         // Should pass with exact minimum
         let result = validate_cluster_fee_from_tags(
@@ -778,7 +779,7 @@ mod tests {
         let minimum_fee = fee_config.compute_fee(
             TransactionType::Hidden,
             tx_size_bytes,
-            effective_wealth,
+            effective_wealth as u128,
             num_memos,
         );
 
@@ -927,8 +928,13 @@ mod tests {
 
         // Calculate the required fee
         let max_wealth = compute_ring_max_cluster_wealth(&ring_tags, &ring_values, &wealth_map);
-        let min_fee =
-            fee_config.compute_fee_with_outputs(TransactionType::Hidden, 4000, max_wealth, 2, 0);
+        let min_fee = fee_config.compute_fee_with_outputs(
+            TransactionType::Hidden,
+            4000,
+            max_wealth as u128,
+            2,
+            0,
+        );
 
         // Should pass with exact minimum
         let result = validate_ring_cluster_fee(
@@ -1087,7 +1093,7 @@ mod tests {
         let min_fee = fee_config.compute_fee_with_dynamic_base_and_outputs(
             TransactionType::Hidden,
             4000,
-            max_wealth,
+            max_wealth as u128,
             2,
             0,
             dynamic_base,
